finish

+++ brief/app/Http/Controllers/ParolesController.php
@@ -2,12 +2,14 @@
 
 namespace App\Http\Controllers;
 use App\Models\Paroless;
+use Auth;
 use Illuminate\Http\Request;
 use App\Http\Resources\ParolesCollection;
 class ParolesController extends Controller{
     //
     public function index(){
         $paroles = Paroless::all();
+        
         return response()->json($paroles);
     }
     public function store(Request $request){
@@ -22,18 +24,32 @@ class ParolesController extends Controller{
     public function show($id){
         return Paroless::find($id);
     }
+
+
+    
     public function update(Request $request, $id){
         $parole =Paroless::find($id);
-        $request->validate([
-            'paroles' => 'required|string|max:255',
-        ]);
-        $parole->paroles = $request->paroles;
-        $parole->user_id = $request->user_id;
-        $parole->musiques_id= $request->musiques_id;
-        $parole->save();
-        return response()->json(['message' => 'Musique updated successfully']);
+        ##dd($parole->user_id);
+     
+        if($parole && $parole->user_id==Auth::user()->id){
+            $request->validate(['paroles' => 'required|string|max:255',]);
+            $parole->paroles = $request->paroles;
+            $parole->user_id = $request->user_id;
+            $parole->musiques_id= $request->musiques_id;
+            $parole->save();
+            return response()->json(['message' => 'paroles updated successfully']);
+        }else{
+            return 'error';
+        }
     }
     public function destroy($id){
-        return Paroless::destroy($id);
+        
+        $parole =Paroless::find($id);
+        if($parole && $parole->user_id==Auth::user()->id){
+            Paroless::destroy($id);
+            return response()->json(['message' => 'paroles delete successfully']);
+        }else{
+            return response()->json(['message'=>'you dont have authoruzation for delete this parole']);
+        }
     }
 }
